// app/types/component.d.ts
export type ButtonVariants = "primary" | "secondary";

export type DashboardSections =
  | "Starter Code"
  | "Discover"
  | "Use Cases"
  | "Enterprise";

export type CodeLanguages =
  | "Python"
  | "Java"
  | "Ruby"
  | "Go"
  | "Typescript"
  | "PHP";

export type UseCaseTitle =
  | "Agents, MCP, & AI Pipelines"
  | "Humans-in-the-Loop"
  | "Compensating Patterns (Saga)"
  | "Long-running Workflows"
  | "Order Fulfillment"
  | "Durable Ledgers"
  | "CI/CD"
  | "Customer Acquisition"
  | "DAG";